+++ backend/funix/decorator/file.py
@@ -10,7 +10,7 @@ from uuid import uuid4
 
 from flask import abort, send_file
 
-from funix import create_safe_tempdir
+from funix.util.file import create_safe_tempdir
 from funix.app import app
 from funix.config.switch import GlobalSwitchOption
 from funix.util.uri import is_valid_uri
